requirements.txt
Django==1.6.2
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
cssmin==0.2.0
distribute==0.6.34
django-apptemplates==0.0.1
django-grappelli==2.5.2
django-model-utils==2.0.3
django-pipeline==1.3.24
django-suit==0.2.7
django-suit-locale==1.0.9
futures==2.1.6
gunicorn==18.0
ply==3.4
psycopg2==2.5.2
python-gettext==2.1
slimit==0.8.1
unittest2==0.5.1
wsgiref==0.1.2
